requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
django-phone-field==1.8.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.4
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
